Update actual_length field in spi_message
Update the actual_length field of the struct spi_message parameter
within the cp2130_spi_transfer_one_message function.  Setting this
field provides reporting of the transaction length which is required
in order for the read/write functions to return the correct length
when the cp2130 driver is used in combination with spidev.

Signed-off-by: David Frey <dfrey@sierrawireless.com>