Zero the bulk transfer buffer before each transfer
In commit c17cced the driver started using a DMA safe memory allocation
for the tx component of the USB bulk transfers.  The problem is that
this buffer must be zero in some reserved fields according to the
datasheet.  Since the reserved fields are never initialized, they are
likely non-zero and this seems to cause the SPI transfers to fail.

Signed-off-by: David Frey <dfrey@sierrawireless.com>